refactor: 경로 검사를 필요로 하는 Piece 분류
