+++ src/main/java/chess/piece/Bishop.java
@@ -4,7 +4,7 @@ import chess.board.Position;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Bishop extends Piece {
+public class Bishop extends SlidingPiece {
 
     private final List<Direction> directions;
 
@@ -33,17 +33,4 @@ public class Bishop extends Piece {
         return new Bishop(positionToMove, this.side);
     }
 
-
-    @Override
-    public List<Position> getPaths(Position targetPosition) {
-        List<Position> paths = new ArrayList<>();
-        final Direction direction = position.getDirectionTo(targetPosition);
-        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
-        Position nextPosition = this.position;
-        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
-            nextPosition = nextPosition.getNextPosition(direction);
-            paths.add(nextPosition);
-        }
-        return paths;
-    }
 }
