+++ src/main/java/chess/piece/Queen.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class Queen extends Piece {
+public class Queen extends SlidingPiece {
 
     private final List<Direction> directions;
     public Queen(final Position position, final Side side) {
@@ -28,16 +28,4 @@ public class Queen extends Piece {
         return new Queen(positionToMove, this.side);
     }
 
-    @Override
-    public List<Position> getPaths(Position targetPosition) {
-        List<Position> paths = new ArrayList<>();
-        final Direction direction = position.getDirectionTo(targetPosition);
-        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
-        Position nextPosition = this.position;
-        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
-            nextPosition = nextPosition.getNextPosition(direction);
-            paths.add(nextPosition);
-        }
-        return paths;
-    }
 }
